move functions that need to be accessed outside of dashboard to viz.py, initial caching implementation
